reorganize apollo-normalized-cache vs apollo-normalized-cache-api
