+++ apollo-normalized-cache-api/README.md
@@ -1,0 +1,3 @@
+# apollo-normalized-cache-api
+
+This is the low-level cache. It knows nothing about coroutines and/or `apollo-runtime`
