+++ apollo-normalized-cache-api/build.gradle.kts
@@ -2,16 +2,23 @@ plugins {
   kotlin("multiplatform")
 }
 
-configureMppDefaults()
+configureMppDefaults(withLinux = false)
 
 kotlin {
   sourceSets {
     val commonMain by getting {
       dependencies {
         api(projects.apolloApi)
+        api(projects.apolloMppUtils)
         implementation(groovy.util.Eval.x(project, "x.dep.okio"))
         api(groovy.util.Eval.x(project, "x.dep.uuid"))
       }
     }
+  }
+}
+
+val jvmJar by tasks.getting(Jar::class) {
+  manifest {
+    attributes("Automatic-Module-Name" to "com.apollographql.apollo3.cache.normalized.api")
   }
 }
