+++ apollo-normalized-cache-api/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/OperationCacheExtensions.kt
@@ -7,7 +7,6 @@ import com.apollographql.apollo3.api.Operation
 import com.apollographql.apollo3.api.json.MapJsonReader
 import com.apollographql.apollo3.api.internal.json.MapJsonWriter
 import com.apollographql.apollo3.api.variables
-import com.apollographql.apollo3.cache.CacheHeaders
 import com.apollographql.apollo3.cache.normalized.internal.CacheBatchReader
 import com.apollographql.apollo3.cache.normalized.internal.Normalizer
