+++ apollo-normalized-cache-api/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/ReadOnlyNormalizedCache.kt
@@ -1,6 +1,5 @@
 package com.apollographql.apollo3.cache.normalized
 
-import com.apollographql.apollo3.cache.CacheHeaders
 import kotlin.jvm.JvmSuppressWildcards
 import kotlin.reflect.KClass
