+++ apollo-normalized-cache-api/src/commonMain/kotlin/com/apollographql/apollo3/cache/normalized/internal/LruCache.kt
