+++ apollo-normalized-cache-api/src/commonTest/kotlin/com/apollographql/apollo3/cache/normalized/internal/LruCacheTest.kt
