--- apollo-normalized-cache-common/build.gradle.kts
@@ -1,18 +1,0 @@
-plugins {
-  kotlin("multiplatform")
-}
-
-configureMppDefaults(withLinux = false)
-
-kotlin {
-  sourceSets {
-    val commonMain by getting {
-      dependencies {
-        api(projects.apolloApi)
-        api(projects.apolloMppUtils)
-        api(projects.apolloNormalizedCacheApi)
-        api(groovy.util.Eval.x(project, "x.dep.kotlin.coroutines"))
-      }
-    }
-  }
-}
