--- apollo-normalized-cache-common/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=apollo-normalized-cache-common
-POM_NAME=Apollo GraphQL Normalized Cache 
-POM_DESCRIPTION=Apollo GraphQL Normalized Cache
