+++ apollo-normalized-cache-sqlite/src/commonTest/kotlin/com/apollographql/apollo3/cache/normalized/sql/SqlNormalizedCacheTest.kt
@@ -1,7 +1,7 @@
 package com.apollographql.apollo3.cache.normalized.sql
 
-import com.apollographql.apollo3.cache.ApolloCacheHeaders
-import com.apollographql.apollo3.cache.CacheHeaders
+import com.apollographql.apollo3.cache.normalized.ApolloCacheHeaders
+import com.apollographql.apollo3.cache.normalized.CacheHeaders
 import com.apollographql.apollo3.cache.normalized.CacheKey
 import com.apollographql.apollo3.cache.normalized.Record
 import kotlin.test.BeforeTest
