+++ apollo-normalized-cache/README.md
@@ -1,0 +1,3 @@
+# apollo-normalized-cache
+
+This module defines `ApolloStore` and `ApolloCacheInterceptor`. `apollo-store` would be a better name but there's just too much stuff out there that uses `apollo-normalized-cache`. Contrary to `apollo-normalized-cache-api`, this module does require
