+++ apollo-normalized-cache/build.gradle.kts
@@ -8,10 +8,8 @@ kotlin {
   sourceSets {
     val commonMain by getting {
       dependencies {
-        api(projects.apolloApi)
-        api(projects.apolloMppUtils)
         api(projects.apolloRuntime)
-        api(projects.apolloNormalizedCacheCommon)
+        api(projects.apolloNormalizedCacheApi)
         api(groovy.util.Eval.x(project, "x.dep.kotlin.coroutines"))
       }
     }
@@ -23,3 +21,4 @@ val jvmJar by tasks.getting(Jar::class) {
     attributes("Automatic-Module-Name" to "com.apollographql.apollo3.cache.normalized")
   }
 }
+
